% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualInvoice.R
\name{manualInvoiceUI}
\alias{manualInvoiceUI}
\title{发票上传生成生成器界面}
\usage{
manualInvoiceUI(
  tabTitle = "电子税务局手工开票信息更新",
  colTitles = c("操作区域", "操作区域", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = manualInvoiceUI_left,
  func_right = manualInvoiceUI_right,
  func_bottom = manualInvoiceUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
发票上传生成生成器界面
}
\examples{
manualInvoiceUI()
}
